requirements.txt
astroid==2.5.6
beautifulsoup4==4.9.3
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.6
filelock==3.0.12
idna==2.10
isort==5.8.0
jieba3k==0.35.1
joblib==1.0.1
lazy-object-proxy==1.6.0
lxml==4.6.3
mccabe==0.6.1
mypy==0.812
mypy-extensions==0.4.3
newspaper3k==0.2.8
nltk==3.6.2
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
pylint==2.8.3
pymongo==3.11.3
python-dateutil==2.8.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
requests-file==1.5.1
sgmllib3k==1.0.0
six==1.15.0
soupsieve==2.2.1
tinysegmenter==0.3
tldextract==3.1.0
toml==0.10.2
tqdm==4.61.1
typed-ast==1.4.3
typing-extensions==3.7.4.3
urllib3==1.26.5
wrapt==1.12.1
